Further code refactor: ImageData now inherits from DataModule, not TabularData.
Given all modules may need a metadata table (including tabular data), I have split it so that DataModule remains abstract but has a metadata table. This has meant the schema structure has had to change to have a metadata and data objects within properties. But has nicely centralised all the code. WriteBinary and fromStream are now called in one place from within DataModule and both TabularData and ImageData have a DataOffset which is now always in the header (it is not module dependent). Previously Imagedata had imageOffset, but Tabular data only had the one table so it did not. Now all modules have metadtaoffset and dataOffset