refactor(ModuleGraph): Optimised reading of Module graph and reduced code duplication
Initially when reading in ModuleGraph from disk, Encounters and Links would be completely read in. Then adjacency and reverese adjacency lists where detected and finally a DFS search of the whole graph was conducted to detect cycles. This required a slightly different DFS search algorithm and meant all links and adjacency lists were completely constructed first. This has been refactored to reuse the wouldCreateCycle method. As each Link is read from file, the adjacency list is updated and wouldCreateCycle is run. This enables early recognitition of cycle detection and prevents an additional run through the graph to construct the adjacency lists.